requirements.txt
nltk==3.8.1
numpy==1.24.2
pandas==1.5.3
requests==2.31.0
scikit-learn==1.2.2
python-dotenv==1.0.0
